+++ hikka/log.py
@@ -13,9 +13,11 @@ import contextlib
 import inspect
 import io
 import json
+import linecache
 import logging
 import os
 import re
+import sys
 import traceback
 import typing
 from logging.handlers import RotatingFileHandler
@@ -25,13 +27,56 @@ import telethon
 from . import utils
 from .tl_cache import CustomTelegramClient
 from .types import BotInlineCall, Module
+from .web.debugger import WebDebugger
+
+# Monkeypatch linecache to make interactive line debugger available
+# in werkzeug web debugger
+# This is weird, but the only adequate approach
+# https://github.com/pallets/werkzeug/blob/3115aa6a6276939f5fd6efa46282e0256ff21f1a/src/werkzeug/debug/tbtools.py#L382-L416
+
+old = linecache.getlines
+
+
+def getlines(filename, module_globals=None):
+    """Get the lines for a Python source file from the cache.
+    Update the cache if it doesn't contain an entry for this file already."""
+
+    try:
+        if filename.startswith("<") and filename.endswith(">"):
+            module = filename[1:-1].split(maxsplit=1)[-1]
+            if (
+                module.startswith("hikka.modules")
+                or module.startswith("dragon.modules")
+            ) and module in sys.modules:
+                return list(
+                    map(
+                        lambda x: f"{x}\n",
+                        sys.modules[module].__loader__.get_source().splitlines(),
+                    )
+                )
+    except Exception:
+        logging.debug("Can't get lines for %s", filename, exc_info=True)
+
+    return old(filename, module_globals)
+
+
+linecache.getlines = getlines
 
 
 class HikkaException:
-    def __init__(self, message: str, local_vars: str, full_stack: str):
+    def __init__(
+        self,
+        message: str,
+        local_vars: str,
+        full_stack: str,
+        sysinfo: typing.Optional[
+            typing.Tuple[object, Exception, traceback.TracebackException]
+        ] = None,
+    ):
         self.message = message
         self.local_vars = local_vars
         self.full_stack = full_stack
+        self.sysinfo = sysinfo
 
     @classmethod
     def from_exc_info(
@@ -88,14 +133,6 @@ class HikkaException:
             ),
             (None, None, None),
         )
-        line = next(
-            (
-                line
-                for line in reversed(full_stack.splitlines())
-                if line.startswith("   ")
-            ),
-            "",
-        )
 
         full_stack = "\n".join(
             [
@@ -122,15 +159,14 @@ class HikkaException:
             message=(
                 f"<b>🚫 Error!</b>\n{cause_mod}\n<b>🗄 Where:</b>"
                 f" <code>{utils.escape_html(filename)}:{lineno}</code><b>"
-                f" in </b><code>{utils.escape_html(name)}</code>\n😵"
-                f" <code>{utils.escape_html(line)}</code>"
-                " 👈\n<b>❓ What:</b>"
+                f" in </b><code>{utils.escape_html(name)}</code>\n<b>❓ What:</b>"
                 f" <code>{utils.escape_html(''.join(traceback.format_exception_only(exc_type, exc_value)).strip())}</code>"
             ),
             local_vars=(
                 f"<code>{utils.escape_html(json.dumps(to_hashable(tb.tb_frame.f_locals), indent=4))}</code>"
             ),
             full_stack=full_stack,
+            sysinfo=(exc_type, exc_value, tb),
         )
 
 
@@ -144,7 +180,7 @@ class TelegramLogsHandler(logging.Handler):
     first trimming handled then unused.
     """
 
-    def __init__(self, targets: list, capacity: int):
+    def __init__(self, targets: list, capacity: int, web_debugger: WebDebugger):
         super().__init__(0)
         self.targets = targets
         self.capacity = capacity
@@ -156,6 +192,8 @@ class TelegramLogsHandler(logging.Handler):
         self._mods = {}
         self.force_send_all = False
         self.tg_level = 20
+        self.web_debugger = web_debugger
+        self._send_lock = asyncio.Lock()
 
     def install_tg_log(self, mod: Module):
         if getattr(self, "_task", False):
@@ -178,7 +216,9 @@ class TelegramLogsHandler(logging.Handler):
         return self.handledbuffer + self.buffer
 
     def dumps(
-        self, lvl: int = 0, client_id: typing.Optional[int] = None
+        self,
+        lvl: int = 0,
+        client_id: typing.Optional[int] = None,
     ) -> typing.List[str]:
         """Return all entries of minimum level as list of strings"""
         return [
@@ -188,7 +228,7 @@ class TelegramLogsHandler(logging.Handler):
             and (not record.hikka_caller or client_id == record.hikka_caller)
         ]
 
-    async def _show_full_stack(
+    async def _show_full_trace(
         self,
         call: BotInlineCall,
         bot: "aiogram.Bot",  # type: ignore
@@ -199,97 +239,130 @@ class TelegramLogsHandler(logging.Handler):
             + "\n\n<b>🦝 Locals:</b>\n"
             + item.local_vars
             + "\n\n"
-            + "<b>🪐 Full stack:</b>\n"
+            + "<b>🪐 Full trace:</b>\n"
             + item.full_stack
         )
 
         chunks = list(utils.smart_split(*telethon.extensions.html.parse(chunks), 4096))
 
-        await call.edit(chunks[0])
+        await call.edit(
+            chunks[0],
+            reply_markup=[
+                {"text": "🐞 Web debugger", "url": item.debug_url},
+            ]
+            if item.debug_url
+            else None,
+        )
 
         for chunk in chunks[1:]:
             await bot.send_message(chat_id=call.chat_id, text=chunk)
 
+    def _gen_web_debug_button(self, item: HikkaException) -> list:
+        if not item.sysinfo:
+            return []
+
+        try:
+            url = self.web_debugger.feed(*item.sysinfo)
+        except Exception:
+            item.debug_url = None
+            return []
+
+        item.debug_url = url
+
+        return [
+            {
+                "text": "🐞 Web debugger",
+                "url": url,
+            }
+        ]
+
     async def sender(self):
-        self._queue = {
-            client_id: utils.chunks(
-                utils.escape_html(
-                    "".join(
-                        [
-                            item[0]
-                            for item in self.tg_buff
-                            if isinstance(item[0], str)
-                            and (
-                                not item[1]
-                                or item[1] == client_id
-                                or self.force_send_all
-                            )
-                        ]
-                    )
-                ),
-                4096,
-            )
-            for client_id in self._mods
-        }
-
-        self._exc_queue = {
-            client_id: [
-                self._mods[client_id].inline.bot.send_message(
-                    self._mods[client_id]._logchat,
-                    item[0].message,
-                    reply_markup=self._mods[client_id].inline.generate_markup(
-                        {
-                            "text": "🪐 Full stack",
-                            "callback": self._show_full_stack,
-                            "args": (
-                                self._mods[client_id].inline.bot,
-                                item[0],
-                            ),
-                            "disable_security": True,
-                        }
-                    ),
-                )
-                for item in self.tg_buff
-                if isinstance(item[0], HikkaException)
-                and (not item[1] or item[1] == client_id or self.force_send_all)
-            ]
-            for client_id in self._mods
-        }
-
-        for exceptions in self._exc_queue.values():
-            for exc in exceptions:
-                await exc
-
-        self.tg_buff = []
-
-        for client_id in self._mods:
-            if client_id not in self._queue:
-                continue
-
-            if len(self._queue[client_id]) > 5:
-                logfile = io.BytesIO("".join(self._queue[client_id]).encode("utf-8"))
-                logfile.name = "hikka-logs.txt"
-                logfile.seek(0)
-                await self._mods[client_id].inline.bot.send_document(
-                    self._mods[client_id]._logchat,
-                    logfile,
-                    caption=(
-                        "<b>🧳 Journals are too big to be sent as separate messages</b>"
-                    ),
-                )
-
-                self._queue[client_id] = []
-                continue
-
-            while self._queue[client_id]:
-                if chunk := self._queue[client_id].pop(0):
-                    asyncio.ensure_future(
-                        self._mods[client_id].inline.bot.send_message(
-                            self._mods[client_id]._logchat,
-                            f"<code>{chunk}</code>",
-                            disable_notification=True,
+        async with self._send_lock:
+            self._queue = {
+                client_id: utils.chunks(
+                    utils.escape_html(
+                        "".join(
+                            [
+                                item[0]
+                                for item in self.tg_buff
+                                if isinstance(item[0], str)
+                                and (
+                                    not item[1]
+                                    or item[1] == client_id
+                                    or self.force_send_all
+                                )
+                            ]
                         )
+                    ),
+                    4096,
+                )
+                for client_id in self._mods
+            }
+
+            self._exc_queue = {
+                client_id: [
+                    self._mods[client_id].inline.bot.send_message(
+                        self._mods[client_id]._logchat,
+                        item[0].message,
+                        reply_markup=self._mods[client_id].inline.generate_markup(
+                            [
+                                {
+                                    "text": "🪐 Full trace",
+                                    "callback": self._show_full_trace,
+                                    "args": (
+                                        self._mods[client_id].inline.bot,
+                                        item[0],
+                                    ),
+                                    "disable_security": True,
+                                },
+                                *self._gen_web_debug_button(item[0]),
+                            ],
+                        ),
                     )
+                    for item in self.tg_buff
+                    if isinstance(item[0], HikkaException)
+                    and (not item[1] or item[1] == client_id or self.force_send_all)
+                ]
+                for client_id in self._mods
+            }
+
+            for exceptions in self._exc_queue.values():
+                for exc in exceptions:
+                    await exc
+
+            self.tg_buff = []
+
+            for client_id in self._mods:
+                if client_id not in self._queue:
+                    continue
+
+                if len(self._queue[client_id]) > 5:
+                    logfile = io.BytesIO(
+                        "".join(self._queue[client_id]).encode("utf-8")
+                    )
+                    logfile.name = "hikka-logs.txt"
+                    logfile.seek(0)
+                    await self._mods[client_id].inline.bot.send_document(
+                        self._mods[client_id]._logchat,
+                        logfile,
+                        caption=(
+                            "<b>🧳 Journals are too big to be sent as separate"
+                            " messages</b>"
+                        ),
+                    )
+
+                    self._queue[client_id] = []
+                    continue
+
+                while self._queue[client_id]:
+                    if chunk := self._queue[client_id].pop(0):
+                        asyncio.ensure_future(
+                            self._mods[client_id].inline.bot.send_message(
+                                self._mods[client_id]._logchat,
+                                f"<code>{chunk}</code>",
+                                disable_notification=True,
+                            )
+                        )
 
     def emit(self, record: logging.LogRecord):
         try:
@@ -387,8 +460,9 @@ def init():
     handler.setLevel(logging.INFO)
     handler.setFormatter(_main_formatter)
     logging.getLogger().handlers = []
+    web_debugger = WebDebugger()
     logging.getLogger().addHandler(
-        TelegramLogsHandler((handler, rotating_handler), 7000)
+        TelegramLogsHandler((handler, rotating_handler), 7000, web_debugger)
     )
     logging.getLogger().setLevel(logging.NOTSET)
     logging.getLogger("telethon").setLevel(logging.WARNING)
@@ -396,4 +470,5 @@ def init():
     logging.getLogger("aiohttp").setLevel(logging.WARNING)
     logging.getLogger("aiogram").setLevel(logging.WARNING)
     logging.getLogger("pyrogram").setLevel(logging.WARNING)
+    logging.getLogger("werkzeug").setLevel(logging.WARNING)
     logging.captureWarnings(True)